requirements.txt
# Jarvis Personal Assistant — Python Dependencies
# Install with: pip install -r requirements.txt --break-system-packages

# ── LLM communication ─────────────────────────────────────────────────────────
httpx>=0.27.0
pydantic>=2.0.0

numpy>=1.24.0

# ── Web search & news — query-driven, no API key, no hardcoded sources ─────────
duckduckgo_search>=6.0.0

# ── Mobile server (WiFi access from phone browser) ────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-telegram-bot>=21.0

# ── Calendar support (optional — for proactive .ics event reminders) ──────────
icalendar>=5.0.0

# ── GTK4 bindings (system package — do NOT install via pip) ───────────────────
# PyGObject  ← sudo pacman -S python-gobject